✨ Feat: 팀페이지 - 제미나이 리뷰 일부 반영
