+++ src/app/(service)/[teamid]/components/TeamTaskList.tsx
@@ -1,32 +1,9 @@
 import TaskGroup from '@/app/(service)/[teamid]/components/TaskGroup';
 import { STATUS } from '@/app/(service)/[teamid]/constants';
-import {
-  TaskItemDetailProps,
-  TaskList,
-  TeamDetailData,
-} from '@/app/(service)/[teamid]/types';
+import { TeamDetailData } from '@/app/(service)/[teamid]/types';
+import { classifyTaskLists } from '@/app/(service)/[teamid]/utils/task';
 
 export default function TeamTaskList({ taskLists }: TeamDetailData) {
-  const getTaskListStatus = (
-    tasks: TaskItemDetailProps[],
-  ): (typeof STATUS)[number] => {
-    if (tasks.length === 0) return '시작 전';
-    const doneCount = tasks.filter((task) => task.doneAt !== null).length;
-    if (doneCount === tasks.length) return '완료';
-    if (doneCount > 0) return '진행 중';
-    return '시작 전';
-  };
-  const classifyTaskLists = (taskLists: TaskList[]) => {
-    const result: Record<(typeof STATUS)[number], TaskList[]> = {
-      '시작 전': [],
-      '진행 중': [],
-      완료: [],
-    };
-    taskLists.forEach((taskList) => {
-      result[getTaskListStatus(taskList.tasks)].push(taskList);
-    });
-    return result;
-  };
   const classified = classifyTaskLists(taskLists);
 
   return (
