+++ src/app/(service)/[teamid]/utils/task.ts
@@ -1,0 +1,24 @@
+import { STATUS } from '@/app/(service)/[teamid]/constants';
+import { TaskItemDetailProps, TaskList } from '@/app/(service)/[teamid]/types';
+
+export const getTaskListStatus = (
+  tasks: TaskItemDetailProps[],
+): (typeof STATUS)[number] => {
+  if (tasks.length === 0) return '시작 전';
+  const doneCount = tasks.filter((task) => task.doneAt !== null).length;
+  if (doneCount === tasks.length) return '완료';
+  if (doneCount > 0) return '진행 중';
+  return '시작 전';
+};
+
+export const classifyTaskLists = (taskLists: TaskList[]) => {
+  const result: Record<(typeof STATUS)[number], TaskList[]> = {
+    '시작 전': [],
+    '진행 중': [],
+    완료: [],
+  };
+  taskLists.forEach((taskList) => {
+    result[getTaskListStatus(taskList.tasks)].push(taskList);
+  });
+  return result;
+};
